common.sh:  when test ends and killing bgs also log a msg about it
this is to be able to tell if for some reason we got leftover bgs or not
and if we get stuck on wait command or not.
also hide kill msgs when waiting for bgs.

+++ common.sh
@@ -1294,8 +1294,12 @@ function fail_if_err() {
 }
 
 function kill_all_bgs() {
-    kill -9 $(jobs -p) 2>/dev/null
-    wait
+    local bgs=$(jobs -p)
+    if [ -n "$bgs" ]; then
+        kill -9 $(jobs -p) 2>/dev/null
+        kmsg "Wait for bgs"
+        wait $bgs &>/dev/null
+    fi
 }
 
 function test_done() {
